🐣 ADD: 유효성 검사 및 sign up 기능 구현
